requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
discord.py==1.3.4
future==0.18.2
idna==2.10
inflect==4.1.0
jaraco.itertools==5.0.0
more-itertools==8.4.0
multidict==4.7.6
oauthlib==3.1.0
praw==7.1.0
prawcore==1.4.0
pycparser==2.20
PyNaCl==1.3.0
python-twitter==3.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
websockets==8.1
wolframalpha==4.0.0
xmltodict==0.12.0
yarl==1.4.2
ccxt==1.46.22
tabulate